requirements.txt
numpy
opencv-python
streamlit
Pillow